Move files in dashboard directory.
