+++ assets/js/modules/reader-revenue-manager/components/dashboard/index.js
@@ -17,3 +17,4 @@
  */
 
 export { default as PublicationApprovedOverlayNotification } from './PublicationApprovedOverlayNotification';
+export { default as ReaderRevenueManagerSetupCTABanner } from './ReaderRevenueManagerSetupCTABanner';
